DJ_API/detector/models/image_model.py
import os
import numpy as np
import tensorflow as tf
from tensorflow.keras.applications.xception import preprocess_input
from PIL import Image
import logging
from django.conf import settings
from datetime import datetime
import time

logger = logging.getLogger(__name__)

# Global variables for persistence
USING_MOCK_MODEL = False
_LOADED_MODEL = None
_MODEL_LOADING_LOCK = False

def load_image_model():
    """Load the TensorFlow image model for deepfake detection"""
    global USING_MOCK_MODEL, _LOADED_MODEL, _MODEL_LOADING_LOCK
    
    # If model is already loaded and persistence is enabled, return it
    if os.environ.get('FORCE_MODEL_PERSISTENCE') == 'true' and _LOADED_MODEL is not None:
        logger.info("Using previously loaded image model (persistence enabled)")
        return _LOADED_MODEL
    
    # Prevent concurrent loading
    if _MODEL_LOADING_LOCK:
        logger.warning("Model loading already in progress, waiting...")
        # Wait for the other thread to complete loading
        while _MODEL_LOADING_LOCK:
            time.sleep(0.1)
        if _LOADED_MODEL is not None:
            return _LOADED_MODEL
    
    _MODEL_LOADING_LOCK = True
    
    try:
        model_path = settings.AI_MODELS['IMAGE_MODEL_PATH']
        use_mock = settings.AI_MODELS.get('USE_MOCK_MODELS', False)
        
        # Force clear any cached setting
        if hasattr(load_image_model, 'use_mock_override'):
            use_mock = load_image_model.use_mock_override
            
        # Check if we should use mock models
        if use_mock:
            from .mock_models import get_mock_image_model
            logger.warning("Using mock image model as specified in settings")
            USING_MOCK_MODEL = True
            _LOADED_MODEL = get_mock_image_model()
            return _LOADED_MODEL
        
        # Check if the model file exists
        if not os.path.exists(model_path):
            logger.error(f"Image model not found at {model_path}")
            logger.error(f"Current directory: {os.getcwd()}")
            logger.error(f"Directory contents: {os.listdir(os.path.dirname(model_path) if os.path.dirname(model_path) else '.')}")
            
            # Check if we have a simple model
            simple_model_path = model_path + ".simple"
            if os.path.exists(simple_model_path):
                logger.info(f"Found simple model at {simple_model_path}, using it instead")
                model_path = simple_model_path
            else:
                # Fall back to mock model
                from .mock_models import get_mock_image_model
                logger.warning("Falling back to mock image model because real model file not found")
                USING_MOCK_MODEL = True
                _LOADED_MODEL = get_mock_image_model()
                return _LOADED_MODEL
        
        logger.info(f"Loading real image model from {model_path}")
        logger.info(f"Model file size: {os.path.getsize(model_path) / (1024 * 1024):.2f} MB")
        
        # Create a simple wrapper class for the model
        class SimpleModelWrapper:
            def __init__(self, model_path):
                self.model_path = model_path
                self.model = None
                self._load_model()
                
            def _load_model(self):
                try:
                    # Try different loading approaches
                    try:
                        # First try with custom objects
                        def add_positional_embedding(x):
                            return x  # Simplified version that just returns input
                        
                        custom_objects = {
                            'add_positional_embedding': add_positional_embedding
                        }
                        
                        self.model = tf.keras.models.load_model(self.model_path, custom_objects=custom_objects)
                        logger.info("Model loaded with custom objects")
                    except Exception as e:
                        logger.error(f"Error loading with custom objects: {str(e)}")
                        
                        # Try without custom objects
                        try:
                            self.model = tf.keras.models.load_model(self.model_path)
                            logger.info("Model loaded without custom objects")
                        except Exception as e2:
                            logger.error(f"Error loading without custom objects: {str(e2)}")
                            
                            # Create a simple model as fallback
                            logger.warning("Creating a simple model as fallback")
                            inputs = tf.keras.Input(shape=(224, 224, 3))
                            x = tf.keras.layers.Conv2D(32, 3, activation="relu")(inputs)
                            x = tf.keras.layers.GlobalAveragePooling2D()(x)
                            outputs = tf.keras.layers.Dense(1, activation="sigmoid")(x)
                            self.model = tf.keras.Model(inputs, outputs)
                            logger.info("Simple fallback model created")
                except Exception as e:
                    logger.error(f"All model loading approaches failed: {str(e)}")
                    raise
            
            def __call__(self, x):
                try:
                    return self.model.predict(x, verbose=0)
                except Exception as e:
                    logger.error(f"Error during prediction: {str(e)}")
                    # Return a fallback prediction
                    return np.array([[0.5]])
        
        # Create the wrapper
        model_wrapper = SimpleModelWrapper(model_path)
        
        # Test with dummy input
        dummy_input = np.zeros((1, 224, 224, 3), dtype=np.float32)
        test_prediction = model_wrapper(dummy_input)
        logger.info(f"Test prediction shape: {test_prediction.shape}")
        logger.info(f"Test prediction value: {test_prediction}")
        
        USING_MOCK_MODEL = False
        _LOADED_MODEL = model_wrapper
        return _LOADED_MODEL
    except Exception as e:
        logger.error(f"Failed to load image model: {str(e)}")
        import traceback
        logger.error(f"Traceback: {traceback.format_exc()}")
        
        # Fall back to mock model
        from .mock_models import get_mock_image_model
        logger.warning("Falling back to mock image model due to error")
        USING_MOCK_MODEL = True
        _LOADED_MODEL = get_mock_image_model()
        return _LOADED_MODEL
    finally:
        _MODEL_LOADING_LOCK = False

# Method to force use of real models
def force_real_models():
    load_image_model.use_mock_override = False
    logger.info("Forced use of real image models")

# Method to force use of mock models
def force_mock_models():
    load_image_model.use_mock_override = True
    logger.info("Forced use of mock image models")

def preprocess_image(image_path, target_size=(224, 224)):
    """Preprocess an image for the model"""
    try:
        start_time = datetime.now()
        logger.info(f"Preprocessing image: {image_path}")
        
        img = Image.open(image_path)
        img = img.convert('RGB')  # Ensure image is RGB
        
        logger.info(f"Original image size: {img.size}")
        
        img = img.resize(target_size)
        img_array = np.array(img)
        
        # Normalize to [0, 1]
        img_array = img_array / 255.0
        
        # Add batch dimension
        img_array = np.expand_dims(img_array, axis=0)
        
        end_time = datetime.now()
        logger.info(f"Image preprocessing completed in {(end_time - start_time).total_seconds():.2f} seconds")
        
        return img_array
    except Exception as e:
        logger.error(f"Error preprocessing image: {str(e)}")
        raise

def predict_image(model, image_path):
    """Make a prediction on an image"""
    global USING_MOCK_MODEL
    
    try:
        start_time = datetime.now()
        logger.info(f"Starting prediction for image: {image_path}")
        logger.info(f"Using mock model: {USING_MOCK_MODEL}")
        
        # Preprocess the image
        processed_img = preprocess_image(image_path)
        
        # Make prediction
        logger.info("Running model inference...")
        prediction_start = datetime.now()
        prediction = model(processed_img)
        
        # Handle different return types (numpy array or tensor)
        if hasattr(prediction, 'numpy'):
            prediction = prediction.numpy()[0][0]
        elif isinstance(prediction, np.ndarray):
            prediction = prediction[0][0]
        else:
            # For mock models or other return types
            prediction = float(prediction[0][0])
            
        prediction_end = datetime.now()
        
        logger.info(f"Raw prediction value: {prediction}")
        logger.info(f"Inference time: {(prediction_end - prediction_start).total_seconds():.4f} seconds")
        
        # Calculate percentages
        fake_percentage = float(prediction) * 100
        real_percentage = 100 - fake_percentage
        
        # Determine if it's a deepfake based on threshold
        is_deepfake = fake_percentage > 50
        
        # Calculate confidence
        confidence = max(fake_percentage, real_percentage)
        
        result = {
            'isDeepfake': bool(is_deepfake),
            'realPercentage': float(real_percentage),
            'fakePercentage': float(fake_percentage),
            'confidence': float(confidence),
            'usingMockModel': USING_MOCK_MODEL
        }
        
        end_time = datetime.now()
        logger.info(f"Prediction completed in {(end_time - start_time).total_seconds():.2f} seconds")
        logger.info(f"Result: {'FAKE' if is_deepfake else 'REAL'} with {confidence:.2f}% confidence")
        
        return result
    except Exception as e:
        logger.error(f"Error during image prediction: {str(e)}")
        logger.exception("Exception details:")
        raise
