DJ_API/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
Pillow==10.1.0
numpy==1.24.3
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.1.78
gunicorn==21.2.0
